requirements.txt
streamlit==1.28.0
JSON==2.0.9
pandas==2.1.0
pillow==9.5.0
matplotlib==3.7.2
seaborn==0.13.0
geoPandas==0.14.0
plotly==5.17.0